Fix building issues on newer Debian and Ubuntu platforms
This fixes the follow error message when trying to build on Debian 12 or Ubuntu
22.04:

  configure: error: cannot find required auxiliary files: config.guess config.sub

I suspect this is due to the fact that this gem uses an ancient version of
gecode which hasn't been updated to work on newer platforms and we're now
running into issues after all these years.

The build commands will update the configure and Makefile scripts so that it
works properly on the platform it's building on. The other changes in
configure.ac are needed to also make those command work without any issues.

Signed-off-by: Lance Albertson <lance@osuosl.org>